Reformat status page
